Transfer some functions from Tracklist.tsx to utils.ts
